Remove support for `--output-format markdown`
It seems to be used very little, if at all. Yet, it significantly slows
down development and maintenance of this project.

So remove it, along with the relevant abstractions. We can always add
the code back later if we end up wanting to support some other output
format.

I have made sure that the use cases for the removed tests have
sufficient coverage elsewhere.
